+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 + Improve whitelist and blacklist helper methods to make them more intuitive. ListenerWrapper.java now has two methods that will tell you which guild has which plugins able to run, and which guilds a plugin can run in, based off the whitelisting rules.
 + Add convenience method that resolves an IMentionable Object from JDA when supplied with an ID (saves typing out code trying to figure out what an ID from a config file represents)
 + Improve AMGN.runCommand logging to show command execution status, and add similar logging for generic command invocations
++ AMGN.runCommand will now respect network whitelist/blacklist rules